Move "pg2dtypes" to utils
